image/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.0
transformers[torch]==4.45.2 # Example version, include the [torch] to minimize extra dependencies
pillow
requests
boto3
urllib3